fix(xt.common): stop ConvergenceStudy::run() from running off the word vector
The new test_convergence_study segfaulted in CI. Its QuadraticStudy reports the
quantity "time to solution", which is longer than a column and is therefore the
first thing in the repository to reach run()'s header word-wrapping branch.

That branch erased the consumed word inside a loop which also advanced the
iterator itself. erase() invalidates the iterator and already yields the one
past the erased element, so the extra ++ skipped a word; once the last word had
been erased the iterator stepped past end(), from where the loop condition never
held again and the iteration walked off the vector. Whether that dropped a word
silently or crashed depended entirely on what happened to lie behind the
allocation -- it survived under the distro toolchain and segfaulted under
clang22.

Assign erase()'s result and only advance when nothing was consumed. The header
now keeps every word ("time to" / "solution" instead of "time" / "solution").

Also add a test which pins the wrapped header, so the branch is covered by an
assertion rather than only being executed, and give the set_prefix logging test
a directory of its own via the ScopedTestDir/CurrentPathGuard pair, so its two
logfiles are removed even when an assertion aborts it early. CurrentPathGuard
moves from configuration.cc into the shared scoped_test_dir.hh for that.
